refactor(shm): carve out core/ submodule for pure-infra primitives
Step 1 of SHM business-decoupling. Move the two truly business-agnostic
modules (PointSlot/seqlock + bitmap allocator) under a `core/` namespace
that becomes the contract: anything inside must remain free of
voltage-model / voltage-routing deps so it can be promoted to a
standalone voltage-shm-slots crate later.

  vec_impl.rs    -> core/slot.rs   (PointSlot, seqlock primitives)
  slot_bitmap.rs -> core/bitmap.rs (SlotBitmap allocator)

Public re-exports preserved; consumers unaffected. Internal references
in channel_index/instance_index/unified_shm updated to crate::core::*.

Other modules (channel_index, instance_index, reverse_index, dispatch,
notifier, unified_shm) still carry business knowledge and stay at the
top level pending Steps 2-5.
