+++ libs/voltage-rtdb-shm/src/core/bitmap.rs
