+++ libs/voltage-rtdb-shm/src/core/slot.rs
