+++ libs/voltage-rtdb-shm/src/instance_index.rs
@@ -34,7 +34,7 @@ use rustc_hash::FxHashMap;
 use std::sync::Arc;
 use voltage_model::PointType;
 
-use crate::slot_bitmap::{SlotAllocation, SlotBitmap};
+use crate::core::bitmap::{SlotAllocation, SlotBitmap};
 
 /// Reference to a shared slot (owned by a Channel)
 #[derive(Clone, Debug, PartialEq)]
